[executorch][serialization] Refactor flatbuffer utils into separate file
For usage in extension/flat_tensor/serialize.

Differential Revision: [D66854756](https://our.internmc.facebook.com/intern/diff/D66854756/)

[ghstack-poisoned]
