+++ exir/_serialize/TARGETS
@@ -33,6 +33,7 @@ runtime.python_library(
         "_dataclass.py",
         "_flatbuffer.py",
         "_program.py",
+        "utils.py",
     ],
     resources = {
         "//executorch/schema:program.fbs": "program.fbs",
